Move lib -> internal per 1.4 changes
